+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+.. _v2-3:
+
+2.3 - `master`_
+~~~~~~~~~~~~~~~
+
+.. note:: This version is not yet released and is under active development.
+
 .. _v2-2:
 
 2.2 - 2018-03-19